Change the logo
I don't think the logo is lining up squarely with the nav links,
and there's just too much going on in these bootstrap styles to
fight with it any longer.

The purpose of a template is to not worry about this stuff. If
we switch to an image logo, then we can revert this commit and
eat cake.